LayoutTests/fast/custom-elements/reactions/CSSStyleDeclaration.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: CEReactions on CSSStyleDeclaration interface</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="cssText, setProperty, setPropertyValue, setPropertyPriority, removeProperty, cssFloat, and all camel cased attributes of CSSStyleDeclaration interface must have CEReactions">
<meta name="help" content="https://dom.spec.whatwg.org/#node">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../imported/w3c/web-platform-tests/custom-elements/resources/custom-elements-helpers.js"></script>
<script src="../../../imported/w3c/web-platform-tests/custom-elements/reactions/resources/reactions.js"></script>
<script src="../resources/additional-helpers.js"></script>
</head>
<body>
<div id="log"></div>
<script>

function test_mutating_style_property_value(testFunction, name, options) {
    const propertyName = (options || {}).propertyName || 'color';
    const idlName = (options || {}).idlName || 'color';
    const value1 = (options || {}).value1 || 'blue';
    const rule1 = `${propertyName}: ${value1};`;
    const value2 = (options || {}).value2 || 'red';
    const rule2 = `${propertyName}: ${value2};`;

    test(function () {
        var element = define_new_custom_element(['style']);
        var instance = document.createElement(element.name);
        assert_array_equals(element.takeLog().types(), ['constructed']);
        testFunction(instance, propertyName, idlName, value1);
        assert_equals(instance.getAttribute('style'), rule1);
        var logEntries = element.takeLog();
        assert_array_equals(logEntries.types(), ['attributeChanged']);
        assert_attribute_log_entry(logEntries.last(), {name: 'style', oldValue: null, newValue: rule1, namespace: null});
    }, name + ' must enqueue an attributeChanged reaction when it adds the observed style attribute');

    test(function () {
        var element = define_new_custom_element(['title']);
        var instance = document.createElement(element.name);
        assert_array_equals(element.takeLog().types(), ['constructed']);
        testFunction(instance, propertyName, idlName, value1);
        assert_equals(instance.getAttribute('style'), rule1);
        assert_array_equals(element.takeLog().types(), []);
    }, name + ' must not enqueue an attributeChanged reaction when it adds the style attribute but the style attribute is not observed');

    test(function () {
        var element = define_new_custom_element(['style']);
        var instance = document.createElement(element.name);
        testFunction(instance, propertyName, idlName, value1);
        assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
        testFunction(instance, propertyName, idlName, value2);
        assert_equals(instance.getAttribute('style'), rule2);
        var logEntries = element.takeLog();
        assert_array_equals(logEntries.types(), ['attributeChanged']);
        assert_attribute_log_entry(logEntries.last(), {name: 'style', oldValue: rule1, newValue: rule2, namespace: null});
    }, name + ' must enqueue an attributeChanged reaction when it mutates the observed style attribute');

    test(function () {
        var element = define_new_custom_element([]);
        var instance = document.createElement(element.name);
        testFunction(instance, propertyName, idlName, value1);
        assert_array_equals(element.takeLog().types(), ['constructed']);
        testFunction(instance, propertyName, idlName, value2);
        assert_equals(instance.getAttribute('style'), rule2);
        assert_array_equals(element.takeLog().types(), []);
    }, name + ' must not enqueue an attributeChanged reaction when it mutates the style attribute but the style attribute is not observed');
}

function test_removing_style_property_value(testFunction, name) {
    test(function () {
        var element = define_new_custom_element(['style']);
        var instance = document.createElement(element.name);
        instance.setAttribute('style', 'color: red; display: none;');
        assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
        testFunction(instance, 'color', 'color');
        assert_equals(instance.getAttribute('style'), 'display: none;'); // Don't make this empty since browser behaviors are inconsistent now.
        var logEntries = element.takeLog();
        assert_array_equals(logEntries.types(), ['attributeChanged']);
        assert_attribute_log_entry(logEntries.last(), {name: 'style', oldValue: 'color: red; display: none;', newValue: 'display: none;', namespace: null});
    }, name + ' must enqueue an attributeChanged reaction when it removes a property from the observed style attribute');

    test(function () {
        var element = define_new_custom_element(['class']);
        var instance = document.createElement(element.name);
        instance.setAttribute('style', 'color: red; display: none;');
        assert_array_equals(element.takeLog().types(), ['constructed']);
        testFunction(instance, 'color', 'color');
        assert_equals(instance.getAttribute('style'), 'display: none;'); // Don't make this empty since browser behaviors are inconsistent now.
        assert_array_equals(element.takeLog().types(), []);
    }, name + ' must not enqueue an attributeChanged reaction when it removes a property from the style attribute but the style attribute is not observed');
}

function test_mutating_style_property_priority(testFunction, name) {
    test(function () {
        var element = define_new_custom_element(['style']);
        var instance = document.createElement(element.name);
        instance.setAttribute('style', 'color: red');
        assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
        testFunction(instance, 'color', 'color', true);
        assert_equals(instance.getAttribute('style'), 'color: red !important;');
        var logEntries = element.takeLog();
        assert_array_equals(logEntries.types(), ['attributeChanged']);
        assert_attribute_log_entry(logEntries.last(), {name: 'style', oldValue: 'color: red', newValue: 'color: red !important;', namespace: null});
    }, name + ' must enqueue an attributeChanged reaction when it makes a property important and the style attribute is observed');

    test(function () {
        var element = define_new_custom_element(['id']);
        var instance = document.createElement(element.name);
        instance.setAttribute('style', 'color: red');
        assert_array_equals(element.takeLog().types(), ['constructed']);
        testFunction(instance, 'color', 'color', true);
        assert_equals(instance.getAttribute('style'), 'color: red !important;');
        assert_array_equals(element.takeLog().types(), []);
    }, name + ' must enqueue an attributeChanged reaction when it makes a property important but the style attribute is not observed');
}

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    instance.style.cssText = `${propertyName}: ${value}`;
}, 'cssText on CSSStyleDeclaration');

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    instance.style.setProperty(propertyName, value);
}, 'setProperty on CSSStyleDeclaration');

test_mutating_style_property_priority(function (instance, propertyName, idlName, isImportant) {
    instance.style.setProperty(propertyName, instance.style[idlName], isImportant ? 'important': '');
}, 'setProperty on CSSStyleDeclaration');

test_removing_style_property_value(function (instance, propertyName, idlName) {
    instance.style.removeProperty(propertyName);
}, 'removeProperty on CSSStyleDeclaration');

test(function () {
    var element = define_new_custom_element(['style']);
    var instance = document.createElement(element.name);
    assert_array_equals(element.takeLog().types(), ['constructed']);
    instance.style.cssFloat = 'left';
    assert_equals(instance.getAttribute('style'), 'float: left;');
    var logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), {name: 'style', oldValue: null, newValue: 'float: left;', namespace: null});
}, 'cssFloat on CSSStyleDeclaration must enqueue an attributeChanged reaction when it adds the observed style attribute');

test(function () {
    var element = define_new_custom_element([]);
    var instance = document.createElement(element.name);
    assert_array_equals(element.takeLog().types(), ['constructed']);
    instance.style.cssFloat = 'left';
    assert_equals(instance.getAttribute('style'), 'float: left;');
    assert_array_equals(element.takeLog().types(), []);
}, 'cssFloat on CSSStyleDeclaration must not enqueue an attributeChanged reaction when it adds the style attribute but the style attribute is not observed');

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    assert_equals(idlName, 'borderWidth');
    instance.style.borderWidth = value;
}, 'A camel case attribute (borderWidth) on CSSStyleDeclaration',
    {propertyName: 'border-width', idlName: 'borderWidth', value1: '2px', value2: '4px'});

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    assert_equals(propertyName, 'border-width');
    instance.style['border-width'] = value;
}, 'A dashed property (border-width) on CSSStyleDeclaration',
    {propertyName: 'border-width', idlName: 'borderWidth', value1: '1px', value2: '5px'});

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    instance.style.webkitFilter = value;
}, 'A webkit prefixed camel case attribute (webkitFilter) on CSSStyleDeclaration',
    {propertyName: 'filter', idlName: 'filter', value1: 'grayscale(20%)', value2: 'grayscale(30%)'});

test_mutating_style_property_value(function (instance, propertyName, idlName, value) {
    instance.style['-webkit-filter'] = value;
}, 'A webkit prefixed dashed property (-webkit-filter) on CSSStyleDeclaration',
    {propertyName: 'filter', idlName: 'filter', value1: 'grayscale(20%)', value2: 'grayscale(30%)'});

</script>
</body>
</html>
